# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /workspaces/{workspaceId}/projects:
    get:
      operationId: listProjects
      tags: [projects]
      summary: List projects in a workspace
      parameters:
        - $ref: "#/components/parameters/workspaceId"
      responses:
        "200":
          description: List of projects
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ProjectSummary"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
    post:
      operationId: createProject
      tags: [projects]
      summary: Create a new project
      parameters:
        - $ref: "#/components/parameters/workspaceId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProjectBody"
      responses:
        "201":
          description: Created project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectSummary"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /projects/{projectId}:
    get:
      operationId: getProject
      tags: [projects]
      summary: Get project with tasks and members
      parameters:
        - $ref: "#/components/parameters/projectId"
      responses:
        "200":
          description: Project details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectDetail"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
    patch:
      operationId: updateProject
      tags: [projects]
      summary: Update a project (team_lead or ADMIN only)
      parameters:
        - $ref: "#/components/parameters/projectId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProjectBody"
      responses:
        "200":
          description: Updated project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectSummary"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      operationId: deleteProject
      tags: [projects]
      summary: Delete a project (team_lead or ADMIN only)
      parameters:
        - $ref: "#/components/parameters/projectId"
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /projects/{projectId}/members:
    get:
      operationId: listProjectMembers
      tags: [projects]
      summary: List project members
      parameters:
        - $ref: "#/components/parameters/projectId"
      responses:
        "200":
          description: List of project members
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ProjectMember"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
    post:
      operationId: addProjectMember
      tags: [projects]
      summary: Add a member to a project (team_lead or ADMIN only)
      parameters:
        - $ref: "#/components/parameters/projectId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddProjectMemberBody"
      responses:
        "201":
          description: Added member
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectMember"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /projects/{projectId}/members/{userId}:
    delete:
      operationId: removeProjectMember
      tags: [projects]
      summary: Remove a member from a project (team_lead or ADMIN only)
      parameters:
        - $ref: "#/components/parameters/projectId"
        - name: userId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Removed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
...
components:
  parameters:
    workspaceId:
      name: workspaceId
      in: path
      required: true
      schema:
        type: string
    projectId:
      name: projectId
      in: path
      required: true
      schema:
        type: string
...
  responses:
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Forbidden:
      description: Forbidden
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    NotFound:
      description: Not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
...
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required: [message]
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]
    User:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        email:
          type: string
        image:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required: [id, name, email, image, createdAt, updatedAt]
...
    ProjectSummary:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
          nullable: true
        priority:
          type: string
          enum: [LOW, MEDIUM, HIGH]
        status:
          type: string
          enum: [ACTIVE, PLANNING, COMPLETED, ON_HOLD, CANCELLED]
        startDate:
          type: string
          format: date-time
          nullable: true
        endDate:
          type: string
          format: date-time
          nullable: true
        teamLead:
          type: string
        workspaceId:
          type: string
        progress:
          type: integer
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        members:
          type: array
          items:
            $ref: "#/components/schemas/ProjectMember"
      required: [id, name, priority, status, teamLead, workspaceId, progress, createdAt, updatedAt, members]
    ProjectDetail:
      allOf:
        - $ref: "#/components/schemas/ProjectSummary"
        - type: object
          properties:
            tasks:
              type: array
              items:
                $ref: "#/components/schemas/TaskDetail"
          required: [tasks]
    ProjectMember:
      type: object
      properties:
        id:
          type: string
        userId:
          type: string
        projectId:
          type: string
        user:
          $ref: "#/components/schemas/User"
        createdAt:
          type: string
          format: date-time
      required: [id, userId, projectId, user, createdAt]
    TaskDetail:
      type: object
      properties:
        id:
          type: string
        projectId:
          type: string
        title:
          type: string
        description:
          type: string
          nullable: true
        status:
          type: string
          enum: [TODO, IN_PROGRESS, DONE]
        type:
          type: string
          enum: [TASK, BUG, FEATURE, IMPROVEMENT, OTHER]
        priority:
          type: string
          enum: [LOW, MEDIUM, HIGH]
        assigneeId:
          type: string
        dueDate:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        assignee:
          $ref: "#/components/schemas/User"
        comments:
          type: array
          items:
            $ref: "#/components/schemas/CommentDetail"
      required: [id, projectId, title, status, type, priority, assigneeId, dueDate, createdAt, updatedAt, assignee, comments]
...
    CommentDetail:
      type: object
      properties:
        id:
          type: string
        content:
          type: string
        userId:
          type: string
        taskId:
          type: string
        createdAt:
          type: string
          format: date-time
        user:
          $ref: "#/components/schemas/User"
      required: [id, content, userId, taskId, createdAt, user]
    CreateProjectBody:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        priority:
          type: string
          enum: [LOW, MEDIUM, HIGH]
        status:
          type: string
          enum: [ACTIVE, PLANNING, COMPLETED, ON_HOLD, CANCELLED]
        startDate:
          type: string
          format: date-time
        endDate:
          type: string
          format: date-time
      required: [name]
    UpdateProjectBody:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        priority:
          type: string
          enum: [LOW, MEDIUM, HIGH]
        status:
          type: string
          enum: [ACTIVE, PLANNING, COMPLETED, ON_HOLD, CANCELLED]
        startDate:
          type: string
          format: date-time
        endDate:
          type: string
          format: date-time
        teamLead:
          type: string
        progress:
          type: integer
          minimum: 0
          maximum: 100
...
    AddProjectMemberBody:
      type: object
      properties:
        userId:
          type: string
      required: [userId]
...