# lib/api-zod/src/generated/types/projectDetail.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ProjectSummary } from "./projectSummary";
import type { TaskDetail } from "./taskDetail";

export type ProjectDetail = ProjectSummary & {
  tasks: TaskDetail[];
};